requirements.txt
altgraph==0.17.4
colorama==0.4.6
dnspython==2.6.1
tqdm==4.66.2
